Split composer package
