+++ src/psd_tools/composer/blend.py
@@ -1,0 +1,29 @@
+"""
+Vector module.
+"""
+from __future__ import absolute_import, unicode_literals
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def blend(target, image, offset):
+    if offset[0] < 0:
+        if image.width <= -offset[0]:
+            return target
+        image = image.crop((-offset[0], 0, image.width, image.height))
+        offset = (0, offset[1])
+
+    if offset[1] < 0:
+        if image.height <= -offset[1]:
+            return target
+        image = image.crop((0, -offset[1], image.width, image.height))
+        offset = (offset[0], 0)
+
+    if target.mode == 'RGBA':
+        target.alpha_composite(image.convert('RGBA'), offset)
+    else:
+        tmp = target.convert('RGBA')
+        tmp.alpha_composite(image.convert('RGBA'), offset)
+        target = tmp.convert(target.mode)
+    return target
